Change pulse count storage from uint64_t to uint32_t
The pulseCnt fields in Emon32Dataset_t and Emon32Cumulative_t are
uint32_t, so align the internal storage and API to match.

Fixes #44

+++ src/pulse.c
@@ -9,7 +9,7 @@
 
 typedef enum PulseLvl_ { PULSE_LVL_LOW, PULSE_LVL_HIGH } PulseLvl_t;
 
-static uint64_t   pulseCount[NUM_OPA];
+static uint32_t   pulseCount[NUM_OPA];
 static PulseCfg_t pulseCfg[NUM_OPA];
 static uint32_t   pinValue[NUM_OPA];
 static PulseLvl_t pulseLvlLast[NUM_OPA];
@@ -44,11 +44,11 @@ void pulseInit(const size_t index) {
   pulseLvlLast[index] = (PulseLvl_t)pinValue[index];
 }
 
-void pulseSetCount(const size_t index, const uint64_t value) {
+void pulseSetCount(const size_t index, const uint32_t value) {
   pulseCount[index] = value;
 }
 
-uint64_t pulseGetCount(const size_t index) { return pulseCount[index]; }
+uint32_t pulseGetCount(const size_t index) { return pulseCount[index]; }
 
 void pulseUpdate(void) {
   uint32_t   mask;
